+++ Controller/UserController.php
@@ -15,13 +15,16 @@ class UserController
         $this->userModel = new UserModel();
     }
 
-    public function isAdmin(): bool {
-
+    public function isAdmin(): bool
+    {
         if (session_status() === PHP_SESSION_NONE) {
             session_start();
         }
 
-        // Vérifie que userId est défini dans la session
+        // Régénération de l'identifiant de session pour éviter la fixation de la session
+        session_regenerate_id(true);
+
+        // Vérifie que l'identifiant de l'utilisateur est défini dans la session
         if (!isset($_SESSION['userId'])) {
             return false;
         }
@@ -37,11 +40,26 @@ class UserController
 
     public function getUserById($id): array
     {
+        // Valide et assainit les données
+        $id = filter_var($id, FILTER_VALIDATE_INT);
+        if ($id === false) {
+            return [];
+        }
+
         return $this->userModel->getByID($id);
     }
 
     public function deleteUser($id): void
     {
+        // Valide et assainit les données
+        $id = filter_var($id, FILTER_VALIDATE_INT);
+        if ($id === false) {
+            // Enregistre l'erreur
+            error_log("Id utilisateur invalide: $id");
+            header("Location: /admin/users");
+            exit();
+        }
+
         $this->userModel->delete($id);
 
         header("Location: /admin/users");
@@ -50,6 +68,15 @@ class UserController
 
     public function changeUserPrivileges($id): void
     {
+        // Valide et assainit les données
+        $id = filter_var($id, FILTER_VALIDATE_INT);
+        if ($id === false) {
+            // Enregistre l'erreur
+            error_log("Id utilisateur invalide: $id");
+            header("Location: /admin/users");
+            exit();
+        }
+
         $this->userModel->toggleAdminStatus($id);
 
         header("Location: /admin/users");
